refactor: convert 함수들 util 로 분리
- api -> display / display -> api 를 해주는 함수들
  types 폴더 말고 utils 폴더에 있도록 수정
